manage Exceptions

+++ HR_Management.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR_Management.Application.Exceptions;
 using HR_Management.Application.Features.LeaveAllocations.Requests.Commands;
 using HR_Management.Application.Persistence.Contracts;
+using HR_Management.Domain;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -20,6 +22,12 @@ namespace HR_Management.Application.Features.LeaveAllocations.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.GetAsync(request.Id);
+
+            if (leaveAllocation == null)
+            {
+                throw new NotFoundException(nameof(LeaveAllocation), request.Id);
+            }
+
             await _leaveAllocationRepository.DeleteAsync(leaveAllocation);
 
             return Unit.Value;
